requirements.txt
Flask
Flask-PyMongo
Werkzeug
python-dotenv
pymongo
dnspython
email-validator
gunicorn
